Fix README about `max_logs` setting
It seems like at the time of the rewrite to TS (7e595e4), this setting
already has a default of 10 (See src/DefaultOptions.ts#L34), probably
means that this is the intended default; but the README about this
setting is not updated.

This commit updates README to reflect this default.